\name{crossReg-package}
\alias{crossReg-package}
\alias{crossReg}
\docType{package}
\title{
Confidence intervals for crossover points 
}
\description{
Given the linear regression model y = b0 + b1*x1 + b2*x2 + b3*x1*x2, the crossover point of the two simple regression lines implied by the linear regression model can be expressed as C = -b2/b3 (Aiken and West, 1991). This package provides functions to calculate confidence intervals for crossover points of two simple linear regression lines using 
 the non-linear regression, the delta method, the Fieller method, and the bootstrap methods.
}
\details{
\tabular{ll}{
Package: \tab crossReg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-08\cr
License: \tab GPL-2\cr
}
}
\author{
Sunbok Lee \cr
Maintainer: Sunbok Lee <sunboklee@gmail.com>
}
